+++ src/job_manager/services.py
@@ -1,6 +1,8 @@
 from datetime import datetime
 
+from api.permission_checker import AbstractPermissionService
 from common.services import model_update
+from django.core.exceptions import PermissionDenied
 from django.db import transaction
 
 from job_manager.models import (
@@ -19,27 +21,43 @@ from job_manager.models import (
 
 
 class WorkCenterService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
+
 
     def create(self, name: str, notes: str) -> WorkCenter:
+        # check for permission to create work center
+        if not self.permission_service.check_for_permission("add_workcenter"):
+            raise PermissionDenied()
+
         work_center = WorkCenter.objects.create(name=name, notes=notes)
         work_center.full_clean()
-        work_center.save()
+        work_center.save(user=self.user)
 
         return work_center
 
     def update(self, work_center: WorkCenter, data: dict) -> WorkCenter:
+        # check for permission to update work center
+        if not self.permission_service.check_for_permission("change_workcenter"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "notes",
         ]
 
-        work_center, _ = model_update(instance=work_center, fields=fields, data=data)
+        work_center, _ = model_update(
+            instance=work_center, fields=fields, data=data, user=self.user
+        )
 
         return work_center
 
     def delete(self, work_center: WorkCenter) -> None:
+        # check for permission to delete work center
+        if not self.permission_service.check_for_permission("delete_workcenter"):
+            raise PermissionDenied()
+
         work_center.delete()
 
 
@@ -49,25 +67,40 @@ class WorkCenterService:
 
 
 class TaskTypeService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
 
-    def create(name: str) -> TaskType:
+    def create(self, name: str) -> TaskType:
+        # check for permission to create task type
+        if not self.permission_service.check_for_permission("add_tasktype"):
+            raise PermissionDenied()
+
         task_type = TaskType.objects.create(name=name)
         task_type.full_clean()
-        task_type.save()
+        task_type.save(user=self.user)
 
         return task_type
 
-    def update(task_type: TaskType, data: dict) -> TaskType:
+    def update(self, task_type: TaskType, data: dict) -> TaskType:
+        # check for permission to update task type
+        if not self.permission_service.check_for_permission("change_tasktype"):
+            raise PermissionDenied()
+
         fields = [
             "name",
         ]
 
-        task_type, _ = model_update(instance=task_type, fields=fields, data=data)
+        task_type, _ = model_update(
+            instance=task_type, fields=fields, data=data, user=self.user
+        )
         return task_type
 
-    def delete(task_type: TaskType) -> None:
+    def delete(self, task_type: TaskType) -> None:
+        # check for permission to delete task type
+        if not self.permission_service.check_for_permission("delete_tasktype"):
+            raise PermissionDenied()
+
         task_type.delete()
 
 
@@ -77,8 +110,9 @@ class TaskTypeService:
 
 
 class TaskService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
 
     @transaction.atomic
     def create(
@@ -98,6 +132,10 @@ class TaskService:
         predecessors: list[Task] = None,
         successors: list[Task] = None,
     ) -> Task:
+        # check for permission to create task
+        if not self.permission_service.check_for_permission("add_task"):
+            raise PermissionDenied()
+
         task = Task.objects.create(
             name=name,
             external_id=external_id,
@@ -112,7 +150,7 @@ class TaskService:
         )
 
         task.full_clean()
-        task.save()
+        task.save(user=self.user)
 
         if dependencies:
             task.dependencies.set(dependencies)
@@ -127,6 +165,10 @@ class TaskService:
 
     @transaction.atomic
     def update(self, *, instance: Task, data: dict) -> Task:
+        # check for permission to update task
+        if not self.permission_service.check_for_permission("change_task"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
@@ -144,12 +186,18 @@ class TaskService:
             "successors",
         ]
 
-        task, _ = model_update(instance=instance, fields=fields, data=data)
+        task, _ = model_update(
+            instance=instance, fields=fields, data=data, user=self.user
+        )
 
         return task
 
     @transaction.atomic
     def delete(self, *, task: Task) -> None:
+        # check for permission to delete task
+        if not self.permission_service.check_for_permission("delete_task"):
+            raise PermissionDenied()
+
         task.delete()
 
 
@@ -159,39 +207,56 @@ class TaskService:
 
 
 class JobTypeService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
 
-    def create(name: str, external_id: str = "", notes: str = "") -> JobType:
+    def create(self, name: str, external_id: str = "", notes: str = "") -> JobType:
+        # check for permission to create job type
+        if not self.permission_service.check_for_permission("add_jobtype"):
+            raise PermissionDenied()
+
         job_type = JobType.objects.create(
             name=name, external_id=external_id, notes=notes
         )
         job_type.full_clean()
-        job_type.save()
+        job_type.save(user=self.user)
 
         return job_type
 
-    def update(job_type: JobType, data: dict) -> JobType:
+    def update(self, job_type: JobType, data: dict) -> JobType:
+        # check for permission to update job type
+        if not self.permission_service.check_for_permission("change_jobtype"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
             "notes",
         ]
 
-        job_type, _ = model_update(instance=job_type, fields=fields, data=data)
+        job_type, _ = model_update(
+            instance=job_type, fields=fields, data=data, user=self.user
+        )
 
         return job_type
 
-    def delete(job_type: JobType) -> None:
+    def delete(self, job_type: JobType) -> None:
+        # check for permission to delete job type
+        if not self.permission_service.check_for_permission("delete_jobtype"):
+            raise PermissionDenied()
+
         job_type.delete()
 
 
 class JobService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
 
     @transaction.atomic
     def create(
+        self,
         name: str,
         due_date: datetime,
         job_type: JobType,
@@ -201,6 +266,10 @@ class JobService:
         notes: str = "",
         priority: int = None,
     ) -> Job:
+        # check for permission to create job
+        if not self.permission_service.check_for_permission("add_job"):
+            raise PermissionDenied()
+
         job = Job.objects.create(
             name=name,
             due_date=due_date,
@@ -212,14 +281,18 @@ class JobService:
         )
 
         job.full_clean()
-        job.save()
+        job.save(user=self.user)
 
         job.update_priority(priority)
 
         return job
 
     @transaction.atomic
-    def update(job: Job, data: dict) -> Job:
+    def update(self, job: Job, data: dict) -> Job:
+        # check for permission to update job
+        if not self.permission_service.check_for_permission("change_job"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "due_date",
@@ -231,7 +304,7 @@ class JobService:
             "dependencies",
         ]
 
-        job, _ = model_update(instance=job, fields=fields, data=data)
+        job, _ = model_update(instance=job, fields=fields, data=data, user=self.user)
 
         # update job priority
         if data.get("priority", None):
@@ -240,7 +313,11 @@ class JobService:
         return job
 
     @transaction.atomic
-    def delete(job: Job) -> None:
+    def delete(self, job: Job) -> None:
+        # check for permission to delete job
+        if not self.permission_service.check_for_permission("delete_job"):
+            raise PermissionDenied()
+
         job.delete()
 
 
@@ -250,19 +327,30 @@ class JobService:
 
 
 class DependencyTypeService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
 
-    def create(name: str, external_id: str = "", notes: str = "") -> DependencyType:
+    def create(
+        self, name: str, external_id: str = "", notes: str = ""
+    ) -> DependencyType:
+        # check for permission to create dependency type
+        if not self.permission_service.check_for_permission("add_dependencytype"):
+            raise PermissionDenied()
+
         dependency_type = DependencyType.objects.create(
             name=name, external_id=external_id, notes=notes
         )
         dependency_type.full_clean()
-        dependency_type.save()
+        dependency_type.save(user=self.user)
 
         return dependency_type
 
-    def update(dependency_type: DependencyType, data: dict) -> DependencyType:
+    def update(self, dependency_type: DependencyType, data: dict) -> DependencyType:
+        # check for permission to update dependency type
+        if not self.permission_service.check_for_permission("change_dependencytype"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "external_id",
@@ -270,18 +358,26 @@ class DependencyTypeService:
         ]
 
         dependency_type, _ = model_update(
-            instance=dependency_type, fields=fields, data=data
+            instance=dependency_type,
+            fields=fields,
+            data=data,
+            user=self.user,
         )
 
         return dependency_type
 
-    def delete(dependency_type: DependencyType) -> None:
+    def delete(self, dependency_type: DependencyType) -> None:
+        # check for permission to delete dependency type
+        if not self.permission_service.check_for_permission("delete_dependencytype"):
+            raise PermissionDenied()
+
         dependency_type.delete()
 
 
 class DependencyService:
-    def __init__(self):
-        pass
+    def __init__(self, user) -> None:
+        self.user = user
+        self.permission_service = AbstractPermissionService(user=user)
 
     @transaction.atomic
     def create(
@@ -295,6 +391,10 @@ class DependencyService:
         tasks: list[Task] = None,
         jobs: list[Job] = None,
     ) -> Dependency:
+        # check for permission to create dependency
+        if not self.permission_service.check_for_permission("add_dependency"):
+            raise PermissionDenied()
+
         dependency = Dependency.objects.create(
             name=name,
             dependency_type=dependency_type,
@@ -304,7 +404,7 @@ class DependencyService:
         )
 
         dependency.full_clean()
-        dependency.save()
+        dependency.save(user=self.user)
 
         if tasks:
             dependency.tasks.set(tasks)
@@ -316,6 +416,10 @@ class DependencyService:
 
     @transaction.atomic
     def update(self, *, instance: Dependency, data: dict) -> Dependency:
+        # check for permission to update dependency
+        if not self.permission_service.check_for_permission("change_dependency"):
+            raise PermissionDenied()
+
         fields = [
             "name",
             "dependency_type",
@@ -326,10 +430,16 @@ class DependencyService:
             "jobs",
         ]
 
-        dependency, _ = model_update(instance=instance, fields=fields, data=data)
+        dependency, _ = model_update(
+            instance=instance, fields=fields, data=data, user=self.user
+        )
 
         return dependency
 
     @transaction.atomic
     def delete(self, instance: Dependency) -> None:
+        # check for permission to delete dependency
+        if not self.permission_service.check_for_permission("delete_dependency"):
+            raise PermissionDenied()
+
         instance.delete()
